feat(cash-flow): display total expense/income amount

+++ src/app/pages/incomes/incomes-view/incomes-view.component.ts
@@ -2,6 +2,7 @@ import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { CashFlow } from '@common/models/cash-flow.model';
 import { Store } from '@ngrx/store';
 import { CashFlowActions, CashFlowSelectors } from '@store/cash-flow';
+import { ConfirmationService } from 'primeng/api';
 import { Observable } from 'rxjs';
 
 @Component({
@@ -11,22 +12,34 @@ import { Observable } from 'rxjs';
       (cashFlowSubmitData)="onSubmit($event)"
       (itemToRemoveId)="removeIncome($event)"
       [cashFlowData]="(incomes$ | async)!"
+      [totalCashFlowAmount]="(totalIncomeAmount$ | async)!"
       [isLoading]="(isLoading$ | async)!"
       [isIncomeMode]="true" />
+
+    <p-confirmDialog />
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class IncomesViewComponent {
   private store: Store = inject(Store);
+  private confirmationService: ConfirmationService = inject(ConfirmationService);
 
   public incomes$: Observable<CashFlow[]> = this.store.select(CashFlowSelectors.incomes);
   public isLoading$: Observable<boolean> = this.store.select(CashFlowSelectors.isLoading);
+  public totalIncomeAmount$: Observable<number> = this.store.select(CashFlowSelectors.totalIncomes);
 
   public onSubmit(incomeData: CashFlow): void {
     this.store.dispatch(CashFlowActions.addIncome({ income: incomeData }));
   }
 
   public removeIncome(incomeId: string): void {
-    this.store.dispatch(CashFlowActions.removeIncome({ incomeId }));
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to proceed?',
+      header: 'Confirmation',
+      icon: 'pi pi-exclamation-triangle',
+      accept: (): void => {
+        this.store.dispatch(CashFlowActions.removeIncome({ incomeId }));
+      },
+    });
   }
 }
